- moving Action base class to ownCloudAppShared - moving FileListTableViewController to ownCloudAppShared, decouple big dependencies like the more view and open item handling so it can remain in the app itself
