+++ ownCloud/Client/Actions/Actions+Extensions/DiscardSceneAction.swift
@@ -18,6 +18,7 @@
 
 import UIKit
 import ownCloudSDK
+import ownCloudAppShared
 
 @available(iOS 13.0, *)
 class DiscardSceneAction: Action {
