+++ ownCloudAppShared/Client/User Interface/ClientItemCell.swift
