+++ ownCloudAppShared/Client/User Interface/SortMethod.swift
